Move shared card configuration to trait
